requirements.txt
psycopg2==2.6.1
alabaster==0.7.6
arrow==0.7.0
Babel==2.1.1
boto==2.38.0
click==5.1
colorama==0.3.3
croniter==0.3.8
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.10.2
django-haystack==2.4.1
django-model-utils==2.3.1
django-reversion==2.0.6
django-sphinxdoc==1.4.0
django-storages-redux==1.3
django-toolbelt==0.0.1
djorm-pgarray==1.2
docopt==0.4.0
docutils==0.12
et-xmlfile==1.0.1
gunicorn==19.1.0
jsonpickle==0.9.2
MarkupSafe==0.23
mock==1.3.0
openpyxl==2.3.5
PasteDeploy==1.5.2
patsy==0.4.1
pbr==1.8.1
Pillow==3.0.0
pypdf2==1.25.1
Pygments==2.0.2
python-dateutil==2.5.2
python-http-client==1.2.3
pytz==2016.3
redis==2.10.3
requests==2.8.1
rq==0.5.6
sendgrid==2.2.1
six==1.10.0
smtpapi==0.3.1
snowballstemmer==1.2.0
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
static3==0.6.1
times==0.7
zope.interface==4.1.3
python-docx==0.8.6
django-axes==2.2.0
django-crispy-forms==1.6.1
